+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Upcoming
 - Refactor and migration to Rx of APILayer
 
+## [v2.1.3](https://github.com/MarekKojder/RxSwiftAPI/tree/2.1.3) (2019-11-18)
+- Fixed downloading files with background configuration.
+- Enabled custom identifier of background configuration.
+
 ## [v2.1.2](https://github.com/MarekKojder/RxSwiftAPI/tree/2.1.2) (2019-11-12)
 - Got rid of `inBackground` parameter.